+++ test/main/Datepicker/api-methods.js
@@ -41,7 +41,7 @@ describe('Datepicker - API methods', function () {
       expect(date.getTime(), 'to be', dateValue(2020, 3, 22));
     });
 
-    it('returns a formatted date stirng of selected date if the format is specified', function () {
+    it('returns a formatted date string of selected date if the format is specified', function () {
       expect(dp.getDate('yyyy-mm-dd'), 'to be', '2020-04-22');
     });
 
@@ -57,8 +57,8 @@ describe('Datepicker - API methods', function () {
 
   describe('setDate()', function () {
     it('changes the selected date to given date', function () {
-      const spyChnageEvent = sinon.spy();
-      input.addEventListener('change', spyChnageEvent);
+      const spyChangeEvent = sinon.spy();
+      input.addEventListener('change', spyChangeEvent);
 
       const viewSwitch = getViewSwitch(picker);
       const date = new Date(2019, 11, 23);
@@ -86,8 +86,8 @@ describe('Datepicker - API methods', function () {
 
       // change by api call should not be a trigger of change event
       // (issue #24)
-      expect(spyChnageEvent.called, 'to be false');
-      input.removeEventListener('change', spyChnageEvent);
+      expect(spyChangeEvent.called, 'to be false');
+      input.removeEventListener('change', spyChangeEvent);
 
       // change the view to the selected daye's days view
       // (issue #33)
@@ -124,8 +124,8 @@ describe('Datepicker - API methods', function () {
     });
 
     it('clears the selection if no dates + clear: true option are given', function () {
-      const spyChnageEvent = sinon.spy();
-      input.addEventListener('change', spyChnageEvent);
+      const spyChangeEvent = sinon.spy();
+      input.addEventListener('change', spyChangeEvent);
 
       const viewSwitch = getViewSwitch(picker);
       const today = dateUtils.today();
@@ -144,8 +144,8 @@ describe('Datepicker - API methods', function () {
 
       // change by api call should not be a trigger of change event
       // (issue #24)
-      expect(spyChnageEvent.called, 'to be false');
-      input.removeEventListener('change', spyChnageEvent);
+      expect(spyChangeEvent.called, 'to be false');
+      input.removeEventListener('change', spyChangeEvent);
     });
 
     it('omits updating the picker UI if render option = false', function () {
@@ -187,7 +187,7 @@ describe('Datepicker - API methods', function () {
       expect(isVisible(picker), 'to be false');
     });
 
-    it('cancels the edit when no valied date is passed if revert option = true', function () {
+    it('cancels the edit when no valid date is passed if revert option = true', function () {
       input.value = '2/14';
       dp.setDate(new Date(-1, 0, 1), {revert: true});
 
@@ -326,7 +326,7 @@ describe('Datepicker - API methods', function () {
       expect(cells[19].textContent, 'to be', '14');
     });
 
-    it('notmalizes iput text\'s format', function () {
+    it('notmalizes input text\'s format', function () {
       const date = new Date(2020, 6, 4);
       input.value = '7 4 2020';
 
@@ -348,7 +348,7 @@ describe('Datepicker - API methods', function () {
       expect(isVisible(picker), 'to be false');
     });
 
-    it('cancels the edit when no valied date is enterd if revert option = true', function () {
+    it('cancels the edit when no valid date is enterd if revert option = true', function () {
       input.value = '0/0/0';
       dp.update({revert: true});
 
@@ -419,7 +419,7 @@ describe('Datepicker - API methods', function () {
       expect(dp.getFocusedDate().getTime(), 'to be', dateValue(2021, 5, 4));
     });
 
-    it('returns date stirng of the focused date if a format string is passed', function () {
+    it('returns date string of the focused date if a format string is passed', function () {
       expect(dp.getFocusedDate('m/d/yy'), 'to be', '4/22/20');
 
       dp.setDate('6/4/2021');
@@ -538,8 +538,8 @@ describe('Datepicker - API methods', function () {
 
   describe('refresh()', function () {
     it('refreshes the input element and picker UI to refrect the internal data', function () {
-      const spyChnageEvent = sinon.spy();
-      input.addEventListener('change', spyChnageEvent);
+      const spyChangeEvent = sinon.spy();
+      input.addEventListener('change', spyChangeEvent);
 
       dp.dates = [dateValue(2020, 1, 14)];
       dp.refresh();
@@ -554,8 +554,8 @@ describe('Datepicker - API methods', function () {
 
       // change by api call should not be a trigger of change event
       // (issue #24)
-      expect(spyChnageEvent.called, 'to be false');
-      input.removeEventListener('change', spyChnageEvent);
+      expect(spyChangeEvent.called, 'to be false');
+      input.removeEventListener('change', spyChangeEvent);
     });
 
     it('also changes the view back to the selected date\'s days view', function () {