+++ test/main/options/multidate.js
@@ -93,7 +93,7 @@ describe('options - multi date', function () {
       expect(getCellIndices(cells, '.selected'), 'to equal', [8, 27]);
       expect(getCellIndices(cells, '.focused'), 'to equal', [27]);
 
-      // setting initial dates does not cuase error
+      // setting initial dates does not cause error
       // (issue #51)
       dp.destroy();
       input.value = '02/14/2020,04/22/2020,03/21/2020';
@@ -347,7 +347,7 @@ describe('options - multi date', function () {
   });
 
   describe('dateDelimiter', function () {
-    it('specifies the date delemiter for the input string', function () {
+    it('specifies the date delimiter for the input string', function () {
       const dp = new Datepicker(input, {maxNumberOfDates: 3, dateDelimiter: '|'});
 
       dp.setDate('2/14/2020', '4/22/2020');