+++ CHANGELOG.md
@@ -114,7 +114,7 @@
 
 ### Changes
 
-- Chnage `datepicer.show()` to move the focus to the input field if it's not focused
+- Change `datepicer.show()` to move the focus to the input field if it's not focused
   - Fix #52 — picker shown by calling show() doesn't hide by clicking outside
 
 ## 1.1.3